+++ modules/api/src/main/java/com/eastbarnetschool/ordermatchingengine/api/repository/BalancesRepository.java
@@ -8,7 +8,7 @@ import java.util.UUID;
 
 @Repository
 public interface BalancesRepository {
-    BalanceEntity updateOrCreateBalance(UUID userId, String ticker, Integer balanceDelta, Integer lockedBalancesDelta);
+    BalanceEntity updateOrCreateBalance(UUID userId, String ticker, Long balanceDelta, Long lockedBalancesDelta);
     ArrayList<BalanceEntity> get(UUID userId);
-    Boolean checkIfHasEnoughBalance(UUID userId, String ticker, Integer requiredBalance);
+    Boolean checkIfHasEnoughBalance(UUID userId, String ticker, Long requiredBalance);
 }
